fix for temporary strings being destroyed before being logged
Also fixes the mixing of DirectMessage and normal streaming to a SimpleLogger object (which overwrote the same buffer and then added it again)
Also means less memory usage (and less chance of OOM) as we are not calling substr() anymore for those large cases envisioned.

+++ src/logging.cpp
@@ -40,7 +40,7 @@ Logger *SimpleLogger::logger = nullptr;
 
 // by the default, display logs with level equal or less than logInfo
 enum LogLevel SimpleLogger::logCurrentLevel = logInfo;
-long long SimpleLogger::maxPayloadLogSize  = 10240;
+uint64_t SimpleLogger::maxPayloadLogSize  = 10240;
 
 #ifndef ENABLE_LOG_PERFORMANCE
 // static member initialization
